Handled ignored errors
The changes either add handling for errors or explicitely ignores
an error by assigning it to a blank (_).

Other changes made while checking out unhandled errors:

in icmp.go:c_icmp the code was using a deprecated method of registering
a handler. The new method is by assigning to a struct field.

in program.go:(*ProgramCollector).Run the code had two distinct paths
for default and non-default interval times. The non-default path was
ignoring the error from c.runProgram and not logging the restarting
or the error when it does occur.